remove tests from unit folder
