+++ src/tests/index.js
@@ -1,9 +1,9 @@
 import "./index-test";
 import "./models/user";
 import "./models/profile";
-import "./unit/models/food.spec";
-import "./unit/models/music.spec";
-import "./unit/models/country.spec";
-import "./unit/models/culture.spec";
-import "./unit/models/comment.spec";
-import "./unit/models/newsletter.spec";
+import "./models/food.spec";
+import "./models/music.spec";
+import "./models/country.spec";
+import "./models/culture.spec";
+import "./models/comment.spec";
+import "./models/newsletter.spec";
